llm_api/requirements.txt
mistralai>=1.0.0
typing_extensions>=4.5.0
python-dotenv>=1.0.0